requirements-dev.txt
black==22.1.0
ipython==7.16.3
isort==5.10.1
pre-commit==2.17.0
PyHamcrest==2.0.3
pytest==7.0.1
pytest-asyncio==0.16.0
